refactor(cli): extract settings command module
